need to fix the getPrice method

+++ src/com/company/Drink.java
@@ -2,15 +2,15 @@ package com.company;
 
 public class Drink implements MenuItem {
     private String name;
-    private int price;
-    public Drink(String name, int price){
+    private double price;
+    public Drink(String name, double price){
         this.name=name;
         this.price=price;
     }
     public String getName(){
         return name;
     }
-    public int getPrice(){
+    public double getPrice(){
         return price;
     }
 }
