+++ src/com/company/Sandwich.java
@@ -3,14 +3,16 @@ package com.company;
 public class Sandwich implements MenuItem {
     private String name;
     private double price;
-    public Sandwich(String name, int price){
+    public Sandwich(String name, double price){
         this.name=name;
         this.price=price;
     }
     public String getName(){
+
         return name;
     }
-    public int getPrice(){
+    public double getPrice(){
+
         return price;
     }
 }
